+++ CHANGELOG.md
@@ -1,6 +1,14 @@
-# CHANGLOG
+# Change Log
 
 `mytosis` follows semantic versioning practices and conforms to [this changelog style](http://keepachangelog.com/en/0.3.0/).
+
+## v1.1.0
+### Added
+- Root reference database constructor.
+- Context constructor.
+
+### Changed
+- `trigger` (private API) returns a standard-library promise instead of a Bluebird promise.
 
 ## v1.0.0
 Initial release.